move read/write interfaces to `io`
