+++ src/arti/views/python.py
@@ -2,10 +2,11 @@ import pickle
 from typing import Any, ClassVar
 
 from arti.formats.pickle import Pickle
+from arti.io.core import read, write
 from arti.storage.local import LocalFile
 from arti.types.core import TypeSystem
 from arti.types.python import python
-from arti.views.core import View, read, write
+from arti.views.core import View
 
 
 class Python(View):
