+++ tests/arti/views/test_python.py
@@ -2,8 +2,8 @@ import pickle
 
 from arti.formats.pickle import Pickle
 from arti.internal.utils import named_temporary_file
+from arti.io.core import read, write
 from arti.storage.local import LocalFile
-from arti.views.core import read, write
 from arti.views.python import Python
